rename(ui-design-tool): separate `api` dir into `api` and `models` dirs
