+++ packages/ui-design-tool/components/UIDesignCanvas/types.ts
@@ -1,4 +1,4 @@
-import { CanvasData } from '@/api/Canvas/model';
+import { CanvasData } from '@/models/Canvas/model';
 
 export interface UIDesignCanvasProps extends React.HTMLAttributes<HTMLDivElement> {
   initialData: CanvasData['children'];
