+++ packages/ui-design-tool/models/Artboard/styles.css.ts
