+++ packages/ui-design-tool/utils/model.test.ts
@@ -1,9 +1,9 @@
-import { Artboard, ArtboardArgs, ArtboardData } from '@/api/Artboard/model';
-import { Canvas, CanvasArgs, CanvasData } from '@/api/Canvas/model';
-import { Layer, LayerArgs, LayerData } from '@/api/Layer/model';
-import { ShapeLayer, ShapeLayerArgs, ShapeLayerData, ShapeType } from '@/api/ShapeLayer/model';
-import { TextLayer, TextLayerArgs, TextLayerData } from '@/api/TextLayer/model';
-import { UIRecord, UIRecordArgs, UIRecordData } from '@/api/UIRecord/model';
+import { Artboard, ArtboardArgs, ArtboardData } from '@/models/Artboard/model';
+import { Canvas, CanvasArgs, CanvasData } from '@/models/Canvas/model';
+import { Layer, LayerArgs, LayerData } from '@/models/Layer/model';
+import { ShapeLayer, ShapeLayerArgs, ShapeLayerData, ShapeType } from '@/models/ShapeLayer/model';
+import { TextLayer, TextLayerArgs, TextLayerData } from '@/models/TextLayer/model';
+import { UIRecord, UIRecordArgs, UIRecordData } from '@/models/UIRecord/model';
 import { LayerType, UIRecordType } from '@/types/Identifier';
 import { NumberUnit, StrokeStylePattern } from '@/types/Unit';
 import { cloneDeep } from '@pigyuma/utils';
